requirements.txt
openai-whisper @ git+https://github.com/openai/whisper.git
requests>=2.25.1
pydub>=0.25.1
ffmpeg-python>=0.2.0
python-dotenv>=0.19.0
gtts>=2.2.3
googletrans==4.0.0-rc1 